Reorganized file locations
