+++ src/utils/distortions.py
@@ -1,14 +1,11 @@
 ### Util library for distorting via polynomials
 
-import json
 from typing import List, Tuple, Union
 
 import numpy as np
 from numpy.typing import NDArray
-from .paths import get_data_path
-from .constants import TRAIN_VAL_TEST_SPLIT
-
-__all__ = ['get_distorted_location', 'paramType', 'get_param_encoding', 'encodings_to_params', 'get_param_split', 'distort_radial', 'distort_tangential']
+from ._paths import get_data_path
+from ._constants import TRAIN_VAL_TEST_SPLIT
 
 paramType = Tuple[float,...]
 
@@ -78,54 +75,3 @@ def distort_tangential(X: NDArray, Y: NDArray, P: Tuple[float, float], x0: float
     X_tangential = (X_til + (2 * P[0] * X_til * Y_til + P[1] * (R2 + 2 * X_til **2) )) / x_scale
     Y_tangential = (Y_til + (P[0] * (R2 + 2 * Y_til **2) + 2 * P[1] * X_til * Y_til)) / y_scale
     return X_tangential, Y_tangential
-
-def get_param_encoding(params: tuple[float,...]) -> str:
-    """
-    Generates a string encoding from a set of distortion parameters
-
-    Args:
-        params (tuple[float]): Distortion parameters
-
-    Returns:
-        str: Filename
-    """
-    return str(abs(hash(params)))
-
-def encodings_to_params(encodings: Union[str, List[str]]) -> tuple[Union[paramType, list[paramType], None], Union[paramType, list[paramType], None]]:
-    """
-    Converts a string encoding or a list of string encodings into a set of distortion parameters
-
-    Args:
-        encodings (str | List[str]): Encodings for distortion parameters
-
-    Returns:
-        tuple[K, P]: K and P either a set of parameters or a list of sets of parameters, depending on the input
-    """
-    # Load encoding mapping
-    with open(get_data_path('hash_to_params.json'), 'r') as f:
-        try:
-            hash_to_params = json.loads(f.read())
-        except:
-            hash_to_params = {}
-
-    # Case: encodings is a list
-    if isinstance(encodings, list):
-        K, P = [], []
-        for encoding in encodings:
-            if encoding not in hash_to_params:
-                print(f'Encoding {encoding} not found in hash_to_params.json')
-                continue
-            K.append(hash_to_params[encoding]['K'])
-            P.append(hash_to_params[encoding]['P'])
-
-    # Case: Single encoding
-    if encodings not in hash_to_params:
-        print(f'Encoding {encodings} not found in hash_to_params.json')
-        return None, None
-    return hash_to_params[encodings]['K'], hash_to_params[encodings]['P']
-
-def get_param_split(params: paramType) -> str:
-    return np.random.default_rng(seed=int(get_param_encoding(params))).choice(
-            ['train', 'val', 'test'],
-            p=TRAIN_VAL_TEST_SPLIT,
-        )
